refactored uploads and action items into libs
